Fixed identification type (Not Integer anymore) and other implementations

+++ src/main/java/com/ceub/pi/effycityservice/repository/AreaTematicaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AreaTematicaRepository extends JpaRepository<AreaTematica, Integer> {
+public interface AreaTematicaRepository extends JpaRepository<AreaTematica, Long> {
 }
